Update README.md
For our CS-101 course project, we were put into teams and asked to create a more complicated version of the famous code-breaking game Mastermind. The project’s workload consisted of analysis, design, programming, and presentation. The goal of the project was to make a 6-digit Mastermind game in which the computer plays as the codebreaker and the user as the codemaker using Java. As the game starts, the computer presents the user with a guess and then expects feedback. The user goes on to compare it to the secret code and provide the computer with two numbers as feedback: the number of digits that are correct and in the correct position and the number of digits that are correct but in the wrong position. The computer then provides another guess based on that feedback. At the end, the computer guesses the secret code. 

The project conditions were as follows: 

- The computer needs to play efficiently, which means no random guessing by the algorithm.
- The computer needs to have the ability to detect if the user has provided inconsistent feedback.
- The I/P and O/P need to happen on the console, no U/I needed.

The project required a deep understanding of Java, data structures, algorithms, and pattern recognition. The final product turned out to be a fully functioning 6-digit Mastermind game with an algorithm that can efficiently guess the secret code in 6.375 guesses on average and detect inconsistent feedback provided by the user.